+++ README.md
@@ -2,20 +2,10 @@
 
 This is a local WebUI customized for the WLAN Pi.
 
-# script deployment
+# deployment requirements (work in progress)
 
-```
-cd ~
-wget https://raw.githubusercontent.com/joshschmelzle/wlanpi-webui/main/setup.sh
-sudo chmod +x setup.sh
-sudo ./setup.sh
-```
+`sudo apt install nginx`
 
-Validate by browsing to `http://<WLANPI_IP>`.
-
-# manual deployment
-
-See `DEPLOYMENT.md` for manual installation instructions.
 
 # thanks
