+++ deployment/nginx/nginx.conf
@@ -1,0 +1,24 @@
+user www-data;
+worker_processes 4;
+pid /run/nginx.pid;
+
+events {
+	worker_connections 256;
+	accept_mutex on;
+	multi_accept on;
+}
+
+http {
+	sendfile on;
+	tcp_nopush on;
+	tcp_nodelay on;
+	keepalive_timeout 15;
+	types_hash_max_size 2048;
+
+	include /etc/nginx/mime.types;
+	default_type application/octet-stream;
+
+	gzip on;
+
+	include /etc/nginx/sites-enabled/*;
+}
