+++ deployment/nginx/wlanpi_speedtest.conf
@@ -1,0 +1,20 @@
+server {
+    listen 8080;
+    server_name _;
+
+    root /var/www/speedtest;
+
+    # execute *.php with php-fpm
+    location ~ \.php$ {
+        fastcgi_split_path_info ^(.+\.php)(/.+)$;
+        fastcgi_pass   unix:/var/run/php/php7.3-fpm.sock;
+        fastcgi_index  index.php;
+
+        fastcgi_param  SCRIPT_FILENAME $document_root$fastcgi_script_name;
+        include fastcgi_params;
+    }
+
+
+    access_log off;
+    error_log /var/log/speedtest_error.log;
+}
