+++ deployment/nginx/wlanpi_webui.conf
@@ -1,0 +1,16 @@
+server {
+    listen 80;
+    server_name _;
+
+    access_log /var/log/webui_access.log;
+    error_log /var/log/webui_error.log;
+
+    location / {
+        # forward application requests to the gunicorn server
+        proxy_pass http://unix:/run/gunicorn.sock;
+        proxy_redirect off;
+        proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+    }
+}
