+++ deployment/systemd/wlanpi_webui.socket
@@ -1,0 +1,9 @@
+[Unit]
+Description=gunicorn socket
+
+[Socket]
+ListenStream=/run/gunicorn.sock
+User=wlanpi
+
+[Install]
+WantedBy=sockets.target
