+++ requirements.txt
@@ -1,2 +1,20 @@
-virtualenv==20.0.21
+appdirs==1.4.4
+click==7.1.2
+distlib==0.3.1
+filelock==3.0.12
+Flask==1.1.2
+gunicorn==20.0.4
+importlib-metadata==1.7.0
+itsdangerous==1.1.0
+Jinja2==2.11.2
+MarkupSafe==1.1.1
+packaging==20.4
+pluggy==0.13.1
+py==1.9.0
+pyparsing==2.4.7
+six==1.15.0
+toml==0.10.1
 tox==3.15.2
+virtualenv==20.0.26
+Werkzeug==1.0.1
+zipp==3.1.0
