+++ scripts/reachability.sh
