+++ wlanpi_webui/__version__.py
@@ -1,17 +1,17 @@
-#                    _                   _                   _           _   
-#          __      _| | __ _ _ __  _ __ (_)    __      _____| |__  _   _(_)  
-#          \ \ /\ / / |/ _` | '_ \| '_ \| |____\ \ /\ / / _ \ '_ \| | | | |  
-#           \ V  V /| | (_| | | | | |_) | |_____\ V  V /  __/ |_) | |_| | |  
-#            \_/\_/ |_|\__,_|_| |_| .__/|_|      \_/\_/ \___|_.__/ \__,_|_|  
-#                                 |_|                                        
-                                    
+#            _                   _                   _           _
+#  __      _| | __ _ _ __  _ __ (_)    __      _____| |__  _   _(_)
+#  \ \ /\ / / |/ _` | '_ \| '_ \| |____\ \ /\ / / _ \ '_ \| | | | |
+#   \ V  V /| | (_| | | | | |_) | |_____\ V  V /  __/ |_) | |_| | |
+#    \_/\_/ |_|\__,_|_| |_| .__/|_|      \_/\_/ \___|_.__/ \__,_|_|
+#                         |_|
+
 """ version information for wlanpi-webui """
 
 __title__ = "wlanpi_webui"
-__description__ = "a custom local WebUI made for the WLAN Pi"
+__description__ = "a custom WebUI made for the WLAN Pi"
 __url__ = "https://github.com/joshschmelzle/wlanpi-webui"
 __author__ = "Josh Schmelzle"
 __author_email__ = "josh@joshschmelzle.com"
-__version__ = "1.0.0-a5"
+__version__ = "1.0.0-beta1"
 __status__ = "alpha"
 __license__ = "BSD-3-Clause"
