+++ wlanpi_webui/errors/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint("errors", __name__)
+
+from wlanpi_webui.errors import handlers
