+++ wlanpi_webui/profiler/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint("profiler", __name__)
+
+from wlanpi_webui.profiler import profiler
