+++ wlanpi_webui/speedtest/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint("speedtest", __name__)
+
+from wlanpi_webui.speedtest import speedtest
