+++ wlanpi_webui/templates/errors/404.html
@@ -1,4 +1,4 @@
-{% extends "layout.html" %}
+{% extends "public/layout.html" %}
 {% block content %}
 <div class="uk-container uk-container-large">
 <div class="uk-section">
