+++ wlanpi_webui/templates/errors/500.html
@@ -1,0 +1,8 @@
+{% extends "public/layout.html" %}
+{% block content %}
+<div class="uk-container uk-container-large">
+<div class="uk-section">
+    <p class="uk-text-large"><span uk-icon="icon: warning; ratio: 2"></span> Internal Server Error.</p>
+</div>
+</div>
+{% endblock %}
