+++ wlanpi_webui/templates/public/networkinfo.html
@@ -1,6 +1,6 @@
-{% extends "layout.html" %} {% block content %}
+{% extends "public/layout.html" %} {% block content %}
 <div class="uk-flex uk-padding uk-flex-column">
-<div><b>FPMS</b> scripts</p></div>
+<div><b>FPMS</b> network information scripts</p></div>
 <div class="uk-grid uk-child-width-1-3@m" uk-grid uk-height-match="target: > div > .uk-card">
     <div>   
       <div class="uk-card uk-card-default uk-card-medium uk-card-hover uk-card-body">
