+++ wlanpi_webui/templates/public/profiler.html
@@ -1,4 +1,4 @@
-{% extends "layout.html" %} {% block content %}
+{% extends "public/layout.html" %} {% block content %}
 <div class="uk-flex-auto">  
 <div class="uk-container uk-padding">
 <p><b>profiler</b> is a tool that helps determine the capabilities of wireless clients. this is accomplished by automating analysis of association frames.</p>
