+++ wlanpi_webui/templates/public/speedtest.html
@@ -1,4 +1,4 @@
-{% extends "layout.html" %}
+{% extends "public/layout.html" %}
 {% block content %}
 <div class="uk-flex uk-flex-1 uk-cover-container">
 <iframe class="uk-cover" style="pointer-events: all;" src="{{ iframe }}" height="99%" width="100%"></iframe>
