+++ wlanpi_webui/wsgi.py
@@ -1,0 +1,12 @@
+#!/usr/bin/python3
+# -*- coding: utf-8 -*-
+
+"""
+wlanpi_webui
+~~~~~~~
+a custom local WebUI made for the WLAN Pi
+"""
+
+from wlanpi_webui.__init__ import create_app
+
+app = create_app()
